feat: refactor burgerConstructorSlice in TS
